% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_rstudio.R
\name{launch_rstudio}
\alias{launch_rstudio}
\title{Launch a new project-less RStudio session}
\usage{
launch_rstudio()
}
\description{
Call this function as an addin to launch a new project-less RStudio session
without closing your currently open RStudio project(s).
}
