% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_daytime_theme.R, R/use_nighttime_theme.R
\name{ude_daytime_theme}
\alias{ude_daytime_theme}
\title{Use daytime theme.}
\usage{
ude_daytime_theme()

ude_daytime_theme()
}
\description{
Quickly switch to daytime theme, currently this is Solarized Light.

Quickly switch to nighttime theme, currently this is Solarized Dark.
}
