% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_nighttime_theme.R
\name{use_nighttime_theme}
\alias{use_nighttime_theme}
\title{Use nighttime theme.}
\usage{
use_nighttime_theme()
}
\description{
Quickly switch to nighttime theme, currently this is Solarized Dark.
}
