% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_teach_mode.R
\name{use_teach_mode}
\alias{use_teach_mode}
\title{Use teacher mode.}
\usage{
use_teach_mode()
}
\description{
Use default theme in RStudio for teaching.
}
\details{
Note: When rstudioapi package allows, will also add functionality to change
font and pane arrangement.
}
